frontend/requirements.txt
Django==4.2.7
requests==2.31.0
python-decouple==3.8
Pillow==10.1.0
gunicorn==21.2.0
whitenoise==6.6.0
mysqlclient==2.2.0
